+++ src/main/java/racingcar/domain/Cars.java
@@ -4,7 +4,7 @@
  * @author hj-woo
  * @version 2.0
  * */
-package step3.racingcar.domain;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
