+++ src/main/java/racingcar/ui/ResultView.java
@@ -5,7 +5,7 @@
  * @version 2.0
  * */
 
-package step3.racingcar.ui;
+package racingcar.ui;
 
 import java.util.ArrayList;
 import java.util.Map;
