+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package step3.racingcar.domain;
+package racingcar.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
